cleaned up some code
replaced active Q_ASSERTs with throw and got the top level catch to save all windows before shutting down



git-svn-id: http://fritzing.googlecode.com/svn/trunk@4013 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/mainwindow.cpp
@@ -1047,7 +1047,10 @@ ModelPart* MainWindow::loadBundledPart(const QString &fileName, bool addToBin) {
 	QDir unzipDir(unzipDirPath);
 
 	QList<ModelPart*> mps = moveToPartsFolder(unzipDir,this,addToBin);
-	Q_ASSERT(mps.count()==1); // if this fails, that means that the bundled was wrong
+	if (mps.count()!=1) {
+		// if this fails, that means that the bundled was wrong
+		throw "bundled count was wrong";
+	}
 
 	rmdir(unzipDirPath);
 
@@ -1134,7 +1137,10 @@ QList<ModelPart*> MainWindow::moveToPartsFolder(QDir &unzipDir, MainWindow* mw, 
 	QStringList namefilters;
 	QList<ModelPart*> retval;
 
-	Q_ASSERT(mw);
+	if (mw == NULL) {
+		throw "MainWindow::moveToPartsFolder mainwindow missing";
+	}
+
 	if(mw) {
 		namefilters << ZIP_SVG+"*";
 		foreach(QFileInfo file, unzipDir.entryInfoList(namefilters)) { // svg files
